refactor: change module names to camelCase
