+++ async-race/src/utils/click-handler/removeCar.ts
