+++ async-race/src/view/createModalWin.ts
