+++ async-race/src/view/garagePage.ts
@@ -1,5 +1,5 @@
 import { getCarsAPI } from '../api/api';
-import { createCars } from '../utils/click-handler/create-cars';
+import { createCars } from '../utils/click-handler/createCars';
 import { CARS_ON_PAGE, START_PAGE_GARAGE } from '../utils/constants';
 import { createButton, createElement, createInput } from '../utils/utils';
